requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.1
pika==1.2.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
python-http-client==3.3.7
sendgrid==6.9.7
starkbank-ecdsa==2.0.3
